qemu-script: First round of a QEMU setup
The prepare.sh script creates a bridge and QEMU VMs with Lokomotive,
requiring a lokoctl binary from:
https://github.com/kinvolk/lokomotive/pull/1305
For now the bridge has a large 172.../12 subnetwork within which the
actual subnet is for the static IP addresses. DHCP is used for iPXE,
and the public interface for network access, and a private interface
(here ipvlan) is using static addresses, subject to change.