Add basic installer creator
The installer creator is a utility that takes a configuration file with
modules to be included in the installer (and commands for how to
include them) and creates a tarball we those files as well as the image
to ship the tarball in.

This installer creator also caches its external assets (pointed by a
URL) so development can be more efficient, and by keeping what and how
files are included in the installer, we can more easily maintain the
installer payloads in the future.

So far only external files (pointed by a URL) are supported, but the
idea is that we can create more module types (files, scripts, etc.),
much like the flatpak builder tool supports, soon.