installer: Move the extraction logic into the container image as well
The script that extracted the installer software from the container into
the host was too dependent on the racker.sh script which is supposed to
be shipped in the image (and thus not easily updatable).

Since this logic will likely need get more complex in the future, this
script moves it into the docker image, making the racker.sh script much
simpler, as it just needs to now run the image command's output into a
file and run it.

This is very important so we can keep the changes to the FCL image at a
very minimum.