installer: Separate the racker-run script into a different image
We need to support different versions of the racker installer, otherwise
if the user just gets a newer version of the installer that e.g. brings
an incompatible "lokoctl", they cannot keep using the previous version
for managing their cluster.

This racker-run script will be in charge of extracting the actual racker
installer contents and running them, so it makes sense for it to also be
in charge of updating/pulling the images. Thus, in order to keep the
racker-run script up to date but without it being dependent on a racker
version, this commit separates it into a racker-runner image.