installer: directly install out of the Docker container
This reduces the role of the Docker image to directly install the
payload when run. This requires certain privileges which can be
checked for (and has the side effect that running it out of curiosity
is harmless and won't extract because the privileges are missing).
A "racker" command for the user is introduced to bootstrap the cluster,
update the current racker version, upgrade to the latest racker
version, or get a particular racker version.