bootstrap: cluster name as env var, decoupled from matchbox/dnsmasq
The caller can still generate a cluster name based on the MAC address
or similar, and if not, a default name is used. Under the current
directory, where the cluster config is, a lokoctl-assets subdirectory
for the terraform assets is set up, while the matchbox and dnsmasq
configuration is moved to a global location.