bootstrap: support static IP addrs
In case static IP addresses are used, they should be set up for the
PXE installation process and for the final OS.
Use an additional CLC snippet for static IP addresses, for now only
the most simple case of one static IPv4 address for one MAC address
per node is covered.
For testing purposes the "racker bootstrap" command will generate a
mapping of addresses that we can use in the test setup but it should be
provided by the user.