wizard: Add support for display & value in options
So far we had to give the same value for options that was also used for
displaying the options to the user, i.e. if we wanted to show the user
an option in a more human-readable way, e.g. "Use DHCP" instead of
"dhcp", we had to then use the former as the value that gets also
assigned to the resulting vars.

Thus, this patch allows options to use a display and value types, so the
display is only used when interacting with the user, and the value is
the one that's assigned to the resulting vars.