bootstrap: help dnsmasq with reusing addresses
With quick reboots the address space got exhausted, maybe due to the
initramfs DHCP client using a different ID than the final system's
client or it was using a temporary one that wasn't used again later,
or it could be that the old leases caused dnsmasq not to hand out
new ones because it wasn't in authoritative mode. However, it did not
help trying to set the authoritative mode and ignore the client ID
and instead use the MAC address only. While these are useful settings,
the workaround for now is to hardcode the assignment.