bootstrap: integrated handling of provisioning failures
Sometimes the PXE-based provisioing may fail due to the settings
applied via IPMI being not reliably applied or when a temporary network
failure hits a path that is not robust enough to recover. In this
cases we can offer to retry the provisioning of the failed nodes.
By default, prompt the user to retry or cancel, and also allow to
specify the behavior upfront through racker command line arguments
(not the secondary bootstrap arguments for the cluster setup).