Prefill the public IP addrs editor
The entries for IP address setup per secondary MAC address need to be
configured in the editor. We can prefill the editor similar to "git
commit" or "git rebase -i" does it. Through comments we can give an
introduction about the expected format and we can generate the list of
entries with valid MAC addresses, so that the user only needs to enter
the IP address configuration.