Add support for the provision choice to the bootstrap command
This patch changes how the bootstrap is taking its arguments and using
them for the 3 wizard args we have (provision-type, lokomotive, and
flatcar). This allows users to seamlessly use just the bootstrap command
with a given -provision flag (as well as the flags respective to the
provision type), or no args to go into wizard mode.

It adapted some of the questions in the wizards too, and renamed the
provision-type arg as just provision.