bootstrap: use systemd-resolved for Kubernetes to find all controllers
The control plane should be reachable as long as a controller is online
which means that there should be multiple IP addresses returned by the
domain name resolution, and if one address does not work, the next will
be tried. With multiple address entries in /etc/hosts systemd-resolved
is able to return multpiple addresses to the querying application but
this behavior is missing in the standard libc resolution which can only
return a single entry from /etc/hosts.
On the mangement node it is easy to set /etc/resolv.conf to the stub
resolver on 127.0.0.53 and kubectl will try all entries it returns.
However, on the cluster nodes we have to circumvent a limitation of the
Kubelet and CoreDNS which can't handle 127.0.0.53 in /etc/resolv.conf.
For them, we will add an extra listener on a dummy interface which they
can reach, too, and use the interface's address in /etc/resolv.conf.
The 192.0.2.254/32 address was chosen in the hope of no clashes with
something else on 192.0.2.254 because this /24 is reserved for
documentation and examples.