bootstrap: show provisioning progress and failures
The different stages of provisioning should be displayed to the user to
understand where an error happens and be able to react on it.
For this purpose the state of provisioning is queried by looking at
created marker files, whether the Kubernetes API is reachable, and
which nodes have registered itself to the cluster. This is combined
with the return code of lokoctl (or terraform for plain Flatcar).
An additional sanity check is done at the beginning to see if the BMCs
are correctly configured before even attempting a PXE boot.
The user is pointed to error logs and debugging commands (a new ipmi
helper subcommand "diag" can show the server summary), and the user
gets the choice offered to exclude a problematic server so that the
cluster can still come up. Later on the user can see similar
information in a new "racker status" command that shows what was
provisioned and what not did not work.