wizard: Allow args to be ignored
Even though arguments' flags or prompts can be skipped, we may need them
to be considered as flags but ignored (because the arguments list may
include arguments processed by other programs besides the wizard).

So this patch implements that and allows it by having a new entry
"ignore: true|false" to the argument entries in the configuration.