installer/racker: Fix ordering in the usage function
The usage function was using an associative array in order to directly
check the first arguments without having to repeat them, but this logic
was changed and no longer used, and it was having the side-effect of
arguments not being printed in the declared order.

This patch simplifies the function + fixes the ordering issue.