installer/racker: Do not use getopt for processing args
We have very few arguments that are processed by getopt, and while
getopt is very useful, it's processing the flags using a different
notation than the args-wizard powered commands (which use a single
dash rather than two).
Another issue with getopt was that it was processing arguments like -h
always as its own args, so we had to provide the args for bootstrap as
"bootstrap -- arg1 arg2".

This patch removes the use of getopt and greatly simplifies the argument
related logic in the racker script while allowing bootstrap to be able
to use the arguments without the -- separator.