installer/racker: update local cache of the versioned image on upgrade
When "racker upgrade" pulls and extracts the "latest" image it writes a
version file with the concrete version. This is later used to run that
version of the Docker image but it could be outdated since it wasn't
explicitly pulled again.
Update the local cache to have the latest content of the versioned
image.