docs: firmware updates through a fwupd container
Firmware updates can be done with containers. Since they need to
be specific for the hardware we can't document how to do it in
general but for those vendor devices covered by fwupd we can
document how to run it in a container.
Give an example fwupd Dockerfile and tell how to execute it.