bootstrap: use minimal DHCP lease time
In case the user changes the hardware, previous IP address assignments
get shifted and it is a bit tricky to force a DHCP renewal. A similar
case is when a different internal subnet is chosen.
Use the minimal DHCP lease time that is possible (2 minutes) to ensure
that a reconfiguration is applied quickly. There may still be races
during the 2 minutes but a simple retry should be enough. For other
cases were a longer lease time was used or IPMI static IP addressing
we can document how to force a renewal or switch to DHCP.