bootstrap: capture Docker ipmitool without \r and stdin
When Docker is run with allocating a TTY it will result in \r\n
instead of \n which can break regular output postprocessing.
This was worked around in one case but we can also avoid it by not
allocating a TTY.
Disable the -t flag through the ipmi helper script env var and also
disable reading from stdin which wasn't intended anyway.