ipmi: add ipmi-prometheus dir to racker
This commit updates the config file to also include ipmi-prometheus
directory as part of the racker binary.

This directory will then be extracted along with other files in the
conf.yaml on the host in the location `/opt/bin`.

The directory contains the following files:

prometheus-operator.lokocfg: Configuration for the Lokomotive component
prometheus-operator.

ipmi.yaml: Manifests for deploying the IPMI exporter image and related
objects such as Service, ServiceMonitor etc so that prometheus-operator
can then start scraping the IPMI metrics.

Signed-off-by: Imran Pochi <imran@kinvolk.io>