racker-sim: use IPMI standard port but different loopback address
The IPMI serial console did not work behind the UDP forwarder because
it used a different port and the IPMI protocol embedds the used port
which then mismatches the expected port on the other side.
The same port has to be used on both ipmi_sim and the forwarder on the
DHCP interface. This requires to use port 623 on the host if we want a
standard port and due to the port clash it requires to use different IP
addresses, too.
Use different loopback addresses which each can listen on port 623.
Since listening on port 623 is disallowed for unprivileged users, set
ambient caps to allow the ipmi_sim process to listen on this port.