bootstrap/ipmi: disconnect dangling serial console sessions
Only one serial console can be attached at a time, and if a dangling
session due to SSH connection problems is still running in the
background, it needs to be disconnected first.
Run the "sol deactivate" command before "sol activate" to make sure
that the connection attempt does not fail due to a dangling session.